Zip UDF fast path breaks with non-dense in order elements (#3564)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3564

The fuzzer discovered an issue with the Zip UDF
https://github.com/facebookincubator/velox/issues/3542

I hadn't accounted for the fact that just because all of the arrays are flat and the same size, doesn't mean they all have the same offsets into their elements Vectors.

E.g. with 3 arrays of size 3, the offsets could be 0, 3, 6 or 6, 3, 0 or 0, 4, 7 (if there are unused elements)

When this is the case the fast path added to the Zip UDF produces incorrect results.

The fix is just to validate the offsets are the same in addition to the sizes.

Reviewed By: kagamiori

Differential Revision: D42183237

fbshipit-source-id: 3093ab55d8c848e411772fd3bfdca32b38ce6671