Table writer 7: ID generator for distinct partitions (#3416)
Summary:
Add an ID generator class PartitionIdGenerator that
generates sequential integer IDs for distinct partition
values. It takes partition RowVector as input and returns
the generated IDs for the corresponding rows.

At the moment this PartitionIdGenerator only supports
single partition. The implementation is based on a single
VectorHasher, which the users could assume to generate
sequential consecutive IDs when adding new partitions.

The limit on the number of partitions that could be processed
is enforced by the user specified "maxPartitions".

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3416

Reviewed By: mbasmanova

Differential Revision: D41725059

Pulled By: gggrace14

fbshipit-source-id: 09d80b2d45272c2c5bca0d57d30e6fd835cc9be7