Fix new size calculation code in in HashTable::checkSize() (#3588)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3588

1. New size was calculated using bits::nextPowerOfTwo(newNumDistincts + 1),
which is incorrect and was giving the same result as the given cpacity until
newNumDistincts reached the capacity. That caused a lot more calls (580 vs 24)
to expensive allocateTables() and rehash(). The fix is to call slightly more complex
version of bits::nextPowerOfTwo(stats_.capacity + 1) instead.

2. Introduced HashTableStats structure with four integers that are useful
to monitor the hash table performance. Added numRehashes_ members
and renamed size_ to capacity_.

3. Exporting members of HashTableStats as operator runtime stats from
HashAggregation and HashBuild.

Differential Revision: D42204898

fbshipit-source-id: 943dd975d93e513b9cb3bd7cf4cbe33ae2a5dfa5