Fix in_map stream column id population
Summary:
Populate the stream identifier metadata properly so feature reordering and feature stats aggregation can pick up in_map streams.

This currently causes the prod feature reordering to be slightly worse than the theoretical ceiling because it pushes the in_map stream of the reordered keys to the front of the stripe instead of grouping with the data streams in the back. However, since the reordered in_maps are still together in chunks, the gap should be very small.

Differential Revision: D42177016

fbshipit-source-id: 9eb0b68bd0e9f7f20000c08d1226225dc9328a10