Remove unneeded memory capping logic (#3630)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3630

When the memory stack was first introduced, the plan was to do accounting in the memory pool and use periodic aggregation to mitigate the potential lock contention on the memory hierarchy.

The design has been changed, and we will fully use memory tracker to delegate the accounting and memory capping logic. We also won't allow custom memory caps for intermediate nodes and only at top level pools instead.

This diff gets rid of the unneeded logic and changes call sites for removal of memory cap concept when adding child.

Reviewed By: xiaoxmeng

Differential Revision: D42256726

fbshipit-source-id: 2d33252fb2bddf3b6ee18e0e954b0b4a3609471e