Backfill data stream sizes for feature stats (#3621)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3621

Feature physical sizes cannot be incrementally aggregated as we write batches in the writer. Instead, it has to be aggregated after every flush.

This diff contains the logic to update data stream sizes and taking care of the nullable size semantics with the new api introduced in the previous diff.

Differential Revision: D42132616

fbshipit-source-id: 943c9b263a5b149286c9221a60ab81f36640c397