Move readNulls out of prepareRead for DWRF
The nulls reading logic is different in DWRF and Parquet. The latter may
have multiple data pages with in place page headers in a single RowGroup,
and the nulls are identified by the repetition/definition levels in the
page headers. The nulls reading need to be aligned with data reading for
Parquet across page boundaries, while DWRF always read nulls before
reading any real data. This commit hoists out the nulls reading and
filtering logic from prepareRead(), so that the Parquet readers and
customize and simplify its nulls handling.