Clear promises in ExchangeQueue::closeLocked() (#3748)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3748

Clear promises in ExchangeQueue::closeLocked().
Failure to do so can leave and leaves zombie Tasks - the tasks that
stay indefinitely in memory, because they are referenced by their
Drivers, which are waiting on the promises from the Exchange Queue.

Also adding fix for a race between Task termination and Driver going
off the thread (see the comment in the code).

Reviewed By: amitkdutta

Differential Revision: D42584959

fbshipit-source-id: dbfcb3e0f8d79f2b08df7c4468def38e04ecdecc