Fix compareMaterializedRows to check all rows in two sets are equal (#3725)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3725

There is a bug in compareMaterializedRows(left, right) that compares two result sets left and right. It currently only check that all rows in left can be found in right, but not the other way around. Hence it would miss the cases like left = {1, 1, 2} and right = {1, 2, 3}. This diff fixes this bug by adding the check the other way around.

This fix discovered an error in the MultiFragmentTest.limit unit test (https://github.com/facebookincubator/velox/issues/3726) where the input vector contains nulls every 7 indices, whereas the expected result given in the test that contains null only at index 0 is incorrect. This diff also fixes this bug in the test.

Reviewed By: Yuhta, mbasmanova

Differential Revision: D42551023

fbshipit-source-id: c07b5333a7db27f85541abeef42e26be5de503b3