Fix VectorPool for custom types (#3855)
Summary:
VectorPool used to keep a cache of vectors keyed by TypeKind. Hence, it could
not distinguish between types that have the same TypeKind and could return a
vector of type VARCHAR when requested a vector of type JSON.

This change is to use raw Type pointer as a cache key and limit the supported
types to singleton built-in types like BIGINT and VARCHAR. Decimal, fixed-width
array, complex and custom types are not supported.

In the future, we may add support for custom singleton types by changing the
type registry to know which custom types are singleton types.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3855

Reviewed By: Yuhta

Differential Revision: D42810934

Pulled By: mbasmanova

fbshipit-source-id: b827d88d21aa8a20df79b57ed1220f86a49cbd90