call finalizeNulls recursivley (#3845)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3845

finalizeNulls is called in the vector writer when a null Is committed to
reverse the effect of any written uncommitted data.
This diff:

1- make sure its called on nested maps and arrays.

before this diff, if we have an Array<Array<int>> and the user added some items in the inner array and then committed null in the outer most array. the space reserved by the items added to the inner most array was not  released because finalize null was not called recursively on the inner array.

2- make it the responsibility of the vector writer.

Differential Revision: D42790113

fbshipit-source-id: 5af509b41c6691fcbbed3ac75d84411da2e1b900