Fix propagation of evaluated results from common subexpression
This fixes a bug in the common sub expression implementation where
if the set of already evaluated shared rows is a subset of the
newly requested rows, then only the intersection of these sets gets
copied over. As a concrete example, if the shared set consists of
{1,2,3} rows and the newly requested rows are {2,3,4} then only {4}
is copied over to the result.

Test Plan:
Modifies an existing test case to ensure this code path is exercised.
The existing test did not catch this bug because it went through a
shortcut in moveOrCopyResult() where to avoid copying, if the result
is a nullptr then the result vector is pointed directly to the full
set of shared rows which already had all the computed values.