Avoid std::variant and utilize dynamic dispatch to support all cast types in GenericWriter (#3827)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3827

With this diff a genericWriter can be casted to any type, this diff add a base class to vector writers
with the functions needed to be called on the casted writer all being virtual.
and represent the casted writer in the VectorWriter<Generic> as a pointer to that base class.
this is useful because:
1. Automatically support all types to be casted to.
2. Does not need to generate code and instantiate all possible writers, only those used (being casted to at any point) are considered, and hence fix the compiler memory blow up.

Performance wise i have some diffs coming up that will
evaluate the performance of the generic writer.

This diff also unify tryCastTo semantics to match that of
GenericView; null at type missmatch and throw at
invalid cast due to double casting to two different types.

Reviewed By: kagamiori

Differential Revision: D42724562

fbshipit-source-id: 612fe6bf48b6b0960f33c32d494332cc1935c15b