Fix Map Fast Path Check
Map::apply detects aligned keys and values arrays and uses a fast path
for this. The test is incomplete since it only looks at the starts of
the arrays and not the sizes. The fast path requires all offsets and
sizes to match between keys and values for all rows, not only the
selection in 'rows'. This is because the result will be checked with
canonicalize(), which checks all elements, thus each index that is
valid for keys must also be valid for values.