Support rowNumberRange for SideTableSplitReader. (#3986)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3986

## Background
This is the first diff for support the side table sampling task.

To support side table sampling in Koski, we need the rowNumberRange in the SideTableSplitReader in order to do the merge.

## Solution
In this diff, we add a member variable in Batch called `rowNumberRange`, and fill in the rowNumber information during the DwrfReader. Besides, in order for the SideTableSplitReader to gain that information, several changes in the QueryBuilder and DwioConfig have been made in this diff to pass the enabling flag to the DwrfReader. Also, `Fill` transformer is used in SideTableSplitReader in order to combine different batch together to gain a batch with specified `batchSize`. A rowCounter is added to the `Fill` operator to combine/split the DwrfReader batch to correctly filling in the rowNumberRange information.

Differential Revision: D43104839

fbshipit-source-id: 9641c8631914b8f2c05083452b825920213c52dc