Add PlanFragment::groupedExecutionLeafNodeIds. (#4058)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4058

The new member (unordered_set) contains leaf plan nodes that
need to be executed in the grouped mode.
It is required since some leaf nodes might still run ungrouped execution
in a fragment that runs with grouped execution strategy.

In this case we have for mixed grouped and ungrouped execution (e.g. broadcast
hash join for a large bucketed-on-join-keys probe-side table).

Example plan:
  SELECT FROM
    SELECT FROM deltoid3.usr_mdf_source_node_fast_rid -- grouped execution
    INNER JOIN
    VALUES -- broadcast
  GROUP BY
  ORDER BY

Differential Revision: D43370236

fbshipit-source-id: 2c49dd98b6c7ae9d7ca4a4e3f3bd6720d4d2272c