Introduce FuzzerConnector (#4094)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4094

FuzzerConnector makes it convenient to use data generated by
VetorFuzzer in query plans. The new connectors takes a VectorFuzzer::Options
and seed as part of the table handle, and generates data according to the
outputType specific by the client.
.
FuzzerConnectorSplits only control the number of rows to be generated.
.
Also, adding assertNumRows() to `AssertQueryBuilder` which only checks t
hat the results have the expected type and number of rows (not their actual
content).

Reviewed By: kagamiori

Differential Revision: D43428261

fbshipit-source-id: 7ad5d918c7bdac8f59fbb6ffe5710a29eb27e9bd