Add task reference in partition output iobuf to prevent memory leaking (#4130)
Summary:
Add a reference to the velox task in partition output buffer (folly iobuf)to prevent
the task to be deleted while the output buffer is accessed outside of partitioned
output buffer manager. For example, as we found in Prestissimo staging test, the
task can be cleanup while some of its output buffers are still held by HTTP server
for sending data response.
This PR fixes the issue to let ByteStream object to take a optional releaseFn when
create folly iobuf and set the releaseFn as part of user data to invoke on iobuf
destruction. For now, only the partition output operator needs to set it. The partition
output operator creates a member std::function which just captures a shared task
reference. A unit test is added to reproduce the race condition and verify the fix.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4130

Reviewed By: oerling

Differential Revision: D43583591

Pulled By: xiaoxmeng

fbshipit-source-id: 515553c9e19a2a8ed90a5bcc04b4b1fa50f879be