Fix exception thrown from CastExpr constructor
Summary:
When evaluating `try(cast(...))`, exceptions from the cast-expression are supported
to be masked by the try-expression. However, CastExpr currently throws exceptions in its
constructor if the from- or to-type has a custom operator that does not support this
translation. This diff fixes this bug by remembering the supported custom casting in the
constructor and handle it in CastExpr::evalSpecialForm() later. We populate the error vector
and make a null-constant result vector in this situation.

This diff fixes https://github.com/facebookincubator/velox/issues/4145.

Differential Revision: D43638621

fbshipit-source-id: b39fa596b27c82ed9e9d862b9c87bbfbaa0922c9