Add subfield pruning in reader and HiveConnector (#3949)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3949

Implement subfield pruning to columns of complex types: arrays, maps and structs.  When a query uses only some of the subfields, the engine provides the complete list of required subfields and the connector is free to prune the rest.

* Pruning a struct means populating some of the members with null values.  This is to ensure the result vector matching the output type even some fields are pruned.  If the field is not in output type at all, we just remove that field instead of populating it with nulls.
* Pruning a map means dropping keys not listed in the required subfields.
* Pruning an array means dropping values with indices larger than maximum required index.

Only one level of subfield is supported for pruning for now.

Reviewed By: oerling

Differential Revision: D42984144

fbshipit-source-id: 8c29730b8e24f6948639076823a6083a26350a74