Restrict signatures of Cast in expression fuzzer (#4162)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4162

Expressiion fuzzer assumes all generated expressions are supported and hence
asserts that retry-with-try should never throw. On the other hand, CastExpr does
throw an exception when evaluating unsupported type translations and this
exception cannot be masked by Try expression. So this diff restrict the signatures
of Cast in expression fuzzer to a subset of supported ones for basic types. In a
subsequent diff, we'll make CastTypedExpr maintain a set of supported signatures
and expose them to expression fuzzer.

This diff fixes https://github.com/facebookincubator/velox/issues/4145.

Reviewed By: mbasmanova

Differential Revision: D43686197

fbshipit-source-id: 3cebe7c8b12c40909d578d4992304fdcfd8c89e1