Add printVector helper function (#4187)
Summary:
The new function generates human-readable text representation of the vector's
data placing each value on a separate line and without truncating. It is
intended to produce outputs that are easy to compare to identify differences
between two vectors. The output is optimized for vectors of complex types with
multiple levels of nestedness.

Examples:

- ARRAY(BIGINT): https://gist.github.com/mbasmanova/324dc7d1c57a76741c5a588466b7a2df
- ARRAY(MAP(VARCHAR(), ARRAY(REAL()))): https://gist.github.com/mbasmanova/2a1fc8ae7277a43a54d460106fc4a4db
- something very complex: https://gist.github.com/mbasmanova/1ab0e465c558486fccb32440fb21d7ba

The output also includes a tree of encodings and types:

https://gist.github.com/mbasmanova/e2c499da568037c394e26a5bb90976db

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4187

Reviewed By: xiaoxmeng, benuofi

Differential Revision: D43769538

Pulled By: mbasmanova

fbshipit-source-id: fe3855a3dc8f1e19c26e42bcf62be1d1de7edb29