Do not throw exception when converting expression to filters (#4206)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4206

We are currently throwing exception when trying to convert expression to `Subfield` and `Filter`.  This is functionally correct, but generates a lot of false alarm error log, and potentially affect the performance by collecting stack trace unnecessarily.  This change improves the way of conversion, and stops throwing exceptions whenever possible.

Reviewed By: oerling

Differential Revision: D43860811

fbshipit-source-id: eb2c1ee3f96fa8ad741291db4b6f144c05f72e59