Fix Aggregate::accumulatorAlignmentSize() (#4268)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4268

There is a problem that `Aggregate::accumulatorAlignmentSize` only applies to the offset inside a group, not the group address itself.  This causes `Aggregate::value` returns unaligned address and crashes for example with `int128_t`.

This change fixes this by aligning the whole group memory, including the allocation run, the normalized key, and the group content itself.

Reviewed By: mbasmanova

Differential Revision: D44015635

fbshipit-source-id: b8fdacd01865579dad7a3c44f130df6467f74237