Fix TryExpr to allow exceptions on a subset of rows of constant result vector
Summary:
TryExpr assumes that when the result vector from its subexpression is a constant vector,
the error vector (if exists) must contain exceptions on all rows. This is not true with
`try(coalesce(foo(c0), 1))` when `c0` has NULLs on some rows and `foo`
processes and throws on the rest of rows. In this situation, `coalesce` returns a constant
vector of 1, but the error vector only contains exceptions on rows that `foo` evaluated on.

This diff fixes TryExpr to allow having exceptions only on a subset of rows when the result
vector from subexpression is a constant vector. This fixes https://github.com/facebookincubator/velox/issues/4235.

Reviewed By: mbasmanova

Differential Revision: D43930836

fbshipit-source-id: 02cf919183f6a99467d9220b36bacf0a1eb9815b