Improve Null Propagation Precision
Many expressions were not recognized as null propagating, thus
reducing effectiveness of running on distinct values in
dictionaries. With this change, a a function call to a default null
behavior function is null propagating if the set of columns null
propagating arguments depend on covers the set of arguments non-null
propagating arguments depend on.

For example substr(c, 1,, if (length(c) > 2, length(c) - 1, 0)) is
null propagating but was not recognized as such because the if is not
null propagating. It depends however on the same set of columns 'c'
depends on, hence a null in c makes a null in substr.