Fix processing empty and skipped splits with dynamic filters
It could happen that a split had a reader with no
SelectiveColumnReader. This would get adaptation transferred to it but
the ransfer would not take place. It coud could also happen that the
first split of a scan was empty and then would accept the ScanSpec of
the next split without transferring the adaptation. The adaptation
consists of dynamic filters and filter order.

We therefore remove methods for transferring adaptation between
RowReaders. Instead we always transfer adaptation from the
pre-existing ScanSpec to the new one.

Transfer of adaptation is needed when DataSources are asynchronously
prepared. The tableScan always has one primary DataSource. However,
reader trees are prepared asynchronously in auxiliary
DataSources. These are moved over the primary one with
setFromDataSource.

We note that dynamic filters rely on the ScanSpec in the DataSource
being physically the same ScanSpec as referenced from the reader
tree. Therefore the ScanSpec is passed from the asynchronously
prepared to the primary but the adaptation is first passed from the
primary to the new one.