Fix reader crash when reading into lazy vector (#4255)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4255

When we reading the struct column, sometimes we read them as lazy vector to facilitate late materialization.  In some cases we read them as lazy first, then in a subsequent stripe, the lazy condition is no longer satisfied and we need to read them eagerly.  In such cases the code did not check whether the result vector we write into is lazy or not, and caused crash because of this.  We takes care of such cases in the reader in this change.

Reviewed By: oerling

Differential Revision: D43983549

fbshipit-source-id: ed0f512bde83e2e62b6b0b301a86ea1a2185246f