Add a way to keep RuntimeMetric::count at 1. (#4329)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4329

For some cases we want RuntimeMetric per entity (Operator),
not per event (like load a lazy vector).
This is to understand the overall metric for a single Driver, so
we can quickly identify a potentical cause for slowness (f.e.).
In this diff we add this ability, defaulting to the old way of
increasing 'count' every time we add a value.
We change two counters to keep the 'count' at 1:
"dataSourceWallNanos" and "dataSourceLazyWallNanos".
This will allow us to see if at least one Driver/thread was in
a bottleneck reading the data.

Differential Revision: D44175164

fbshipit-source-id: e9047f5e57e2ff7848e1cb859c3b78780b72e027