remove argTypes and returnType from simple functions metadata (#4278)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4278

metadata for simple  used to expose the return
type and the arg types, however those are not always fixed; For example,
when there are variables in the signature, or when there is variadic.

The metadata now have the signature instead which
can be used to extract such information.

This diff removes returnType and argTypes from the metadata, for some
users that depends on reading the return type from the metadata the diff
add tryResolveReturnType, that may return null if the output is not fixed.

Reviewed By: mbasmanova

Differential Revision: D44045196

fbshipit-source-id: f7a643349846740727b607c878c5cc96d0f51b93