Use applyRow in Expr::applyFunctionWithPeeling()
Summary:
Fuzzer found a bug in Expr::applyFunctionWithPeeling() when evaluating the subscript()
function. The exception throws when creating indices of size `rows.end() == 100`
with an argument vector of size 99. The argument vector was evaluated with non-null
rows (i.e., `applyRows`) up to row 99, so we should make indices with non-null rows
instead of all rows as well. In general, inputs to applyFunctionWithPeeling() is evaluated
with applyRows, so code inside this function should not process rows outside of this
range.

This diff fixes this bug by making applyFunctionWithPeeling() use applyRows instead of
rows. It fixes https://github.com/facebookincubator/velox/issues/4340.

Differential Revision: D44192368

fbshipit-source-id: 716bae60cfb4f2a02fcac91473fe4103ab035c97